# frontend/openapi.json
{
  "components" : {
    "schemas" : {
...
      "ShapesDocumentInfo" : {
        "properties" : {
          "default" : {
            "type" : "boolean"
          },
          "enabled" : {
            "type" : "boolean"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "order" : {
            "format" : "int32",
            "type" : "integer"
          },
          "origin" : {
            "enum" : [
              "IMPORTED",
              "AUTHORED"
            ],
            "type" : "string"
          },
          "sourceFileName" : {
            "type" : "string"
          },
          "tripleCount" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "contact" : {
      "name" : "soptim",
      "url" : "https://www.soptim.de/"
    },
    "description" : "This API provides utilities for editing RDFGraphs that model UML classes using the CIM standard.",
    "license" : {
      "name" : "Apache License 2.0",
      "url" : "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "title" : "RDFArchitect backend",
    "version" : "0.0.0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/datasets/{datasetName}/graphs/{graphURI}/shacl/documents" : {
      "get" : {
        "description" : "Lists the sets of SHACL shapes stored for a graph, in order, without their content.",
        "operationId" : "listShapesDocuments",
        "parameters" : [
          {
            "description" : "The name/url of the inquirer.",
            "in" : "header",
            "name" : "Origin",
            "required" : false,
            "schema" : {
              "default" : "unknown",
              "type" : "string"
            }
          },
          {
            "description" : "The literal name of the dataset.",
            "in" : "path",
            "name" : "datasetName",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "description" : "The url encoded uri of the graph, or \"default\" to access the default graph.",
            "in" : "path",
            "name" : "graphURI",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/ShapesDocumentInfo"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "list constraints documents",
        "tags" : [
          "shacl"
        ]
      },
      "post" : {
        "description" : "Creates a set of SHACL shapes from a Turtle string.",
        "operationId" : "createShapesDocument",
        "parameters" : [
          {
            "description" : "The name/url of the inquirer.",
            "in" : "header",
            "name" : "Origin",
            "required" : false,
            "schema" : {
              "default" : "unknown",
              "type" : "string"
            }
          },
          {
            "description" : "The literal name of the dataset.",
            "in" : "path",
            "name" : "datasetName",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "description" : "The url encoded uri of the graph, or \"default\" to access the default graph.",
            "in" : "path",
            "name" : "graphURI",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "description" : "Display name of the new document.",
            "in" : "query",
            "name" : "name",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "text/plain" : {
              "schema" : {
                "description" : "The SHACL shapes in Turtle syntax.",
                "type" : "string"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ShapesDocumentInfo"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "add a constraints document from Turtle",
        "tags" : [
          "shacl"
        ]
      }
    },
    "/api/datasets/{datasetName}/graphs/{graphURI}/shacl/documents/file" : {
      "post" : {
        "description" : "Uploads a set of SHACL shapes. TTL, RDF/XML and N-Triples are accepted; anything other than Turtle is converted to Turtle on import.",
        "operationId" : "createShapesDocumentFromFile",
        "parameters" : [
          {
            "description" : "The name/url of the inquirer.",
            "in" : "header",
            "name" : "Origin",
            "required" : false,
            "schema" : {
              "default" : "unknown",
              "type" : "string"
            }
          },
          {
            "description" : "The literal name of the dataset.",
            "in" : "path",
            "name" : "datasetName",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "description" : "The url encoded uri of the graph, or \"default\" to access the default graph.",
            "in" : "path",
            "name" : "graphURI",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "description" : "Display name; defaults to the file name.",
            "in" : "query",
            "name" : "name",
            "required" : false,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "properties" : {
                  "file" : {
                    "description" : "The file containing the SHACL shapes.",
                    "format" : "binary",
                    "type" : "string"
                  }
                },
                "required" : [
                  "file"
                ],
                "type" : "object"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ShapesDocumentInfo"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "add a constraints document from a file",
        "tags" : [
          "shacl"
        ]
      }
    },
...
  }
}